Moved ostm encode/decode into separate file and added tests
